feat/#14: 전체 옷장 조회 api 연결
