+++ hooks/queries/allCloset.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 
-import { deleteAllClosetProduct, fetchAllCloset, updateAllClosetProduct } from '@/apis/allCloset';
+import { fetchAllCloset } from '@/apis/allCloset';
 
 const QUERY_KEY = {
   allCloset: 'allCloset',
@@ -10,29 +10,27 @@ const QUERY_KEY = {
 
 export const useFetchAllCloset = () => {
   const { data } = useQuery([QUERY_KEY.allCloset], fetchAllCloset);
-  return {
-    allCloset: data,
-  };
+  return data;
 };
 
 /** Mutation */
 
-export const useUpdateAllClosetProductMutation = () => {
-  const queryClient = useQueryClient();
+// export const useUpdateAllClosetProductMutation = () => {
+//   const queryClient = useQueryClient();
 
-  return useMutation(updateAllClosetProduct, {
-    onSuccess() {
-      queryClient.invalidateQueries([QUERY_KEY.allCloset]);
-    },
-  });
-};
+//   return useMutation(updateAllClosetProduct, {
+//     onSuccess() {
+//       queryClient.invalidateQueries([QUERY_KEY.allCloset]);
+//     },
+//   });
+// };
 
-export const useDeleteAllClosetProductMutation = () => {
-  const queryClient = useQueryClient();
+// export const useDeleteAllClosetProductMutation = () => {
+//   const queryClient = useQueryClient();
 
-  return useMutation(deleteAllClosetProduct, {
-    onSuccess() {
-      queryClient.invalidateQueries([QUERY_KEY.allCloset]);
-    },
-  });
-};
+//   return useMutation(deleteAllClosetProduct, {
+//     onSuccess() {
+//       queryClient.invalidateQueries([QUERY_KEY.allCloset]);
+//     },
+//   });
+// };
